Move armor updates to HumanEntity, don't send armor packet right when joining
